// src/interfaces.ts
export interface IProject {
  thumbnail: string;
  title: string;
  description: string;
  author: string;
  authorPicture: string;
  date: string;
  link: string;
}

export interface IPost {
  type: string;
  title: string;
  description: string;
  author: string;
  authorPicture: string;
  date: string;
  time: string;
}